TipuriExceptieController.php Restricted only for admin

+++ frontend/controllers/TipuriExceptieController.php
@@ -37,13 +37,16 @@ class TipuriExceptieController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new TipuriExceptieSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if(SystemController::userIsAdmin()){
+            $searchModel = new TipuriExceptieSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+        return  $this->redirect(['site/index']);
     }
 
     /**
@@ -54,9 +57,14 @@ class TipuriExceptieController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(SystemController::userIsAdmin()){
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+        return  $this->redirect(['site/index']);
+
+
     }
 
     /**
@@ -66,19 +74,24 @@ class TipuriExceptieController extends Controller
      */
     public function actionCreate()
     {
-        $model = new TipuriExceptie();
+        if(SystemController::userIsAdmin()){
+            $model = new TipuriExceptie();
 
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'id' => $model->id]);
+                }
+            } else {
+                $model->loadDefaultValues();
             }
-        } else {
-            $model->loadDefaultValues();
-        }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        }
+        return  $this->redirect(['site/index']);
+
+
     }
 
     /**
@@ -90,15 +103,20 @@ class TipuriExceptieController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(SystemController::userIsAdmin()){
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
+        return  $this->redirect(['site/index']);
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+
     }
 
     /**
@@ -110,9 +128,12 @@ class TipuriExceptieController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(SystemController::userIsAdmin()){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        return  $this->redirect(['site/index']);
     }
 
     /**
