+++ frontend/controllers/UserController.php
@@ -38,7 +38,7 @@ class UserController extends Controller
      */
     public function actionIndex()
     {
-        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin')){
+        if(SystemController::userIsAdmin()){
             $searchModel = new UserSearch();
             $dataProvider = $searchModel->search($this->request->queryParams);
 
@@ -58,7 +58,7 @@ class UserController extends Controller
      */
     public function actionView($id)
     {
-        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin')){
+        if(SystemController::userIsAdmin()){
             return $this->render('view', [
                 'model' => $this->findModel($id),
             ]);
@@ -73,7 +73,7 @@ class UserController extends Controller
      */
     public function actionCreate()
     {
-        if(!\Yii::$app->user->getIsGuest() && \Yii::$app->user->can('admin')){
+        if(SystemController::userIsAdmin()){
             $model = new User();
 
             if ($this->request->isPost) {
@@ -102,15 +102,18 @@ class UserController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(SystemController::userIsAdmin()){
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+        return  $this->redirect(['site/index']);
     }
 
     /**
@@ -122,9 +125,13 @@ class UserController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(SystemController::userIsAdmin()){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        return  $this->redirect(['site/index']);
+
     }
 
     /**
